Added pagable functionality

+++ Tik_Tok_Spring_Project/src/main/java/com/example/tiktokproject/services/FileService.java
@@ -1,5 +1,6 @@
 package com.example.tiktokproject.services;
 
+import com.example.tiktokproject.exceptions.NotFoundException;
 import org.springframework.stereotype.Service;
 
 import javax.servlet.http.HttpServletResponse;
@@ -13,6 +14,9 @@ public class FileService {
 
     public void download(String filename, HttpServletResponse response) {
         File f = new File("photos" + File.separator + filename);
+        if (!f.exists()) {
+            throw new NotFoundException("File doesn't exist");
+        }
         try {
             Files.copy(f.toPath(), response.getOutputStream());
         } catch (IOException e) {
